Fix compilation error

+++ src/openrct2/rct1/S4Importer.cpp
@@ -2579,7 +2579,7 @@ namespace RCT1
          */
         void SetConvertedWaterRideSpeed()
         {
-            for (size_t rideIndex = 0; rideIndex < _isConvertedWaterRide.capacity(); rideIndex++)
+            for (uint16_t rideIndex = 0; rideIndex < _isConvertedWaterRide.capacity(); rideIndex++)
             {
                 if (_isConvertedWaterRide[rideIndex] && !_hadAChainLiftBefore[rideIndex])
                 {
